REFACTORING : package 이름 소문자 변경 및 config package 생성
