+++ src/main/java/gift/authorization/Token.java
@@ -1,4 +1,4 @@
-package gift.Authorization;
+package gift.authorization;
 
 public record Token(String token) {
 }
