+++ src/main/java/gift/controller/ProductController.java
@@ -1,16 +1,15 @@
 package gift.controller;
 
-import gift.Entity.Product;
+import gift.entity.Product;
 import gift.service.ProductService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import gift.Entity.*;
+
 import java.util.Collection;
 import java.util.Optional;
-import gift.*;
 
 
 @RestController
