+++ src/main/java/gift/entity/User.java
@@ -1,4 +1,4 @@
-package gift.Entity;
+package gift.entity;
 
 import java.beans.ConstructorProperties;
