+++ src/main/java/gift/repository/JdbcUserRepository.java
@@ -1,7 +1,7 @@
 package gift.repository;
 
-import gift.Entity.User;
-import gift.ExceptionHandler.DuplicateValueException;
+import gift.entity.User;
+import gift.exceptionhandler.DuplicateValueException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
