+++ src/main/java/gift/repository/JdbcWishRepository.java
@@ -1,6 +1,6 @@
 package gift.repository;
 
-import gift.Entity.Wish;
+import gift.entity.Wish;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
