+++ src/main/java/gift/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package gift.repository;
 
-import gift.Entity.Product;
+import gift.entity.Product;
 import jakarta.validation.Valid;
 
 import java.util.List;
